stock_sentiment_multiagent/requirements.txt
google-genai>=1.0.0
langgraph>=0.2.0
langchain-core>=0.3.0
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
